Change result value if datatype and language are given
There seems to be a missing condition if a literal contains data type and language information. In this case an empty string is returned. In this commit, a condition is added so that at least the language information is returned. It would be worth considering whether a string with both information should be returned, but since EasyRDF seems to serialize only the language information in this case, it is omitted at this point.